the start of cmdliner.py and cmdliner_json.py
NOTE: WRITTEN USING PYTHON3 ONLY

cmdliner.py is a sqlite3-backed app that provides a silly interface to
adding/removing/running shell commands based on the flask intro
tutorial; It isn't complete and will probably punch/stab any child within a 5
mile radius if ran. Don't do it. And it's dirty-REST.

cmdliner_json.py is a pure REST implementation of the same concept,
however commands are derrived from a datastructure in the .py file. This
implementation is much more functional and the actual desired project
result.  yay.  Need pluggable authentication module to make it useful;
need to expand some of the functionality of the API.